+++ modules/resolver-rule-associations/versions.tf
@@ -1,7 +1,10 @@
 terraform {
-  required_version = ">= 0.13.1"
+  required_version = ">= 1.3.2"
 
   required_providers {
-    aws = ">= 3.56"
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 3.56"
+    }
   }
 }