+++ modules/zones/outputs.tf
@@ -13,7 +13,17 @@ output "route53_zone_name_servers" {
   value       = { for k, v in aws_route53_zone.this : k => v.name_servers }
 }
 
+output "primary_name_server" {
+  description = "The Route 53 name server that created the SOA record."
+  value       = { for k, v in aws_route53_zone.this : k => v.primary_name_server }
+}
+
 output "route53_zone_name" {
   description = "Name of Route53 zone"
   value       = { for k, v in aws_route53_zone.this : k => v.name }
+}
+
+output "route53_static_zone_name" {
+  description = "Name of Route53 zone created statically to avoid invalid count argument error when creating records and zones simmultaneously"
+  value       = { for k, v in var.zones : k => lookup(v, "domain_name", k) if var.create }
 }